refactor: 파일을 폴더로 묶어서 관리한다
