+++ src/container/ListContainer.jsx
@@ -1,8 +1,8 @@
 import { useSelector, useDispatch } from 'react-redux';
 
-import List from './List';
+import List from '../presentational/List';
 
-import { deleteTask } from './actions';
+import { deleteTask } from '../actions';
 
 export default function ListContainer() {
   const { tasks } = useSelector((state) => ({
